1. Se cambiaron los status en el panel de proyectos, 2. Ahora se pueden eliminar financiamientos en el panel de proyectos. 3. Ahora se pueden poner cantidades grandes de numeros en monto, 4. Cambios menores

+++ database/migrations/2016_07_13_050234_create_financings_table.php
@@ -15,7 +15,7 @@ class CreateFinancingsTable extends Migration
         Schema::create('financings', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->decimal('mount', 5, 2);
+            $table->double('mount', 15, 5);
             $table->integer('project_id')->unsigned();
             $table->integer('institution_id')->unsigned();
